feat: deletes data.ts, adds new controller files
